refactor(domains): move favorites-store to domains/catalog (#232)
