+++ src/app/(buyer)/cuenta/favoritos/FavoritosClient.tsx
@@ -7,7 +7,7 @@ import { HeartIcon } from '@heroicons/react/24/solid'
 import { ShoppingCartIcon } from '@heroicons/react/24/outline'
 import { useT } from '@/i18n'
 import { useCartStore } from '@/lib/cart-store'
-import { useFavoritesStore } from '@/lib/favorites-store'
+import { useFavoritesStore } from '@/domains/catalog/favorites-store'
 
 interface FavoriteProduct {
   id: string
