+++ src/components/catalog/FavoriteToggleButton.tsx
@@ -5,7 +5,7 @@ import { useSession } from 'next-auth/react'
 import { useRouter } from 'next/navigation'
 import { HeartIcon as HeartOutline } from '@heroicons/react/24/outline'
 import { HeartIcon as HeartSolid } from '@heroicons/react/24/solid'
-import { useFavoritesStore } from '@/lib/favorites-store'
+import { useFavoritesStore } from '@/domains/catalog/favorites-store'
 import { createAnalyticsItem, trackAnalyticsEvent } from '@/lib/analytics'
 import { useT } from '@/i18n'
 import { cn } from '@/lib/utils'
