+++ src/domains/catalog/favorites-store.ts
